wip big alloc metrics Current is crashing on mod.rs:604 due to an underflow. I believe this is because addresses allocated with MiniPages always started after the MiniPage header. The problem is with big allocation allocated addresses can be anywhere within the (2k + size_of<MiniPage header>()) block of memory, in fact they should always be the first byte of this block. So any big alloc addr will always cause this exact underflow.
A solution could be to pad big allocations to start after this first bit of memory, however this wouldn't always fix the problem. Because a user could try to internally de-allocate the next MiniPage's header address. So it would only be a patchwork solution.

Luckily the upcoming change to stop embedding headers in the user allocatable heap, and put them in a static range at the start, will fix this problem. So for now this crash will remain.